refactor: user로 기능 이전
