--- src/domain/signout/entity/signout.entity.ts
@@ -1,7 +1,0 @@
-import { Provider } from 'src/domain/user/user.entity';
-
-export { Provider };
-
-export interface SignoutResponseEntity {
-  is_updated: boolean;
-}
