+++ src/domain/user/user.result.ts
@@ -1,0 +1,3 @@
+export interface UserResponseResult {
+  is_updated: boolean;
+}
