+++ src/infrastructure/user/user.core-repository.ts
@@ -6,6 +6,7 @@ import { TokenEntity } from 'src/domain/user/token.entity';
 import { SignupCommand } from 'src/domain/auth/command/signup.command';
 import { DomainCustomException } from 'src/domain/common/errors/domain-custom-exception';
 import { DomainErrorCode } from 'src/domain/common/errors/domain-error-code';
+import { UserRequestCommand } from 'src/domain/user/user.command';
 
 @Injectable()
 export class UserCoreRepository implements UserRepository {
@@ -121,5 +122,28 @@ export class UserCoreRepository implements UserRepository {
         DomainErrorCode.DB_SERVER_ERROR
       );
     }
+  }
+
+  async signout(requestDto: UserRequestCommand) {
+    await this.prisma.$transaction(async (prisma) => {
+      const rawData = await prisma.user.updateMany({
+        where: {
+          user_id: requestDto.userId,
+          email: requestDto.email,
+          is_deleted: false,
+        },
+        data: { is_deleted: true, updated_at: new Date() },
+      });
+      if (rawData.count === 0 || rawData.count > 1) {
+        throw new DomainCustomException(500, DomainErrorCode.DB_SERVER_ERROR);
+      }
+    });
+    await this.prisma.token.updateMany({
+      where: {
+        user_id: requestDto.userId,
+      },
+      data: { refresh_token: '', updated_at: new Date() },
+    });
+    return true;
   }
 }
